Slim Autocut PR: Object menu Remove silences + dapi media autocut
Move jump-cut UI from ActionBar to Object menu → AI actions → Remove
silences. Keep shared computeAutocut in runtime, CLI command, reference
doc, and minimal web engine glue. Drop local whisper stack, ActionBar
pill, mock docs, and check-script churn.
