+++ apps/cli/package.json
@@ -8,18 +8,15 @@
   },
   "exports": {
     "./protocol": "./src/protocol.ts",
-    "./channels": "./src/cli-channels.ts",
-    "./local-transcribe": "./src/local-transcribe.ts"
+    "./channels": "./src/cli-channels.ts"
   },
   "scripts": {
     "build": "esbuild src/index.ts --bundle --platform=node --format=cjs --external:esbuild --external:@babel/core --external:@babel/preset-typescript --external:babel-preset-solid --external:bufferutil --external:utf-8-validate --outfile=dist/index.js && chmod +x dist/index.js",
-    "check": "tsc --noEmit && node --experimental-strip-types --test src/local-transcribe.test.ts",
-    "test": "node --experimental-strip-types --test src/local-transcribe.test.ts",
+    "check": "tsc --noEmit",
     "symlink:remove": "rm -f /opt/homebrew/bin/dapi",
     "symlink:create": "npm run build && ln -sf \"$PWD/dist/index.js\" /opt/homebrew/bin/dapi"
   },
   "dependencies": {
-    "@diffusionstudio/runtime": "*",
     "@babel/core": "^7.29.7",
     "@trpc/client": "^11.18.0",
     "@trpc/server": "^11.18.0",
