+++ apps/web/src/components/genai/selection.ts
@@ -36,7 +36,6 @@ export function useMediaSelection() {
 
   const imageNodes = tagged("image");
   const videoNodes = tagged("video");
-  const audioNodes = tagged("audio");
 
   const bound = createMemo(() => {
     const lib = library();
@@ -53,5 +52,5 @@ export function useMediaSelection() {
 
   const images = createMemo(() => bound().filter((entry) => entry.asset.type === "IMAGE"));
 
-  return { bound, images, imageNodes, videoNodes, audioNodes };
+  return { bound, images, imageNodes, videoNodes };
 }
